requirements.txt
antlr4-python3-runtime==4.13.2
asgiref==3.10.0
blinker==1.9.0
click==8.3.0
Django==5.2.7
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
sqlparse==0.5.3
Werkzeug==3.1.3
